+++ wgpu/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 - introduce `ShaderModuleDescriptor`
+- introduce `RenderEncoder`
 
 ### v0.6.2 (2020-11-24)
 - don't panic in the staging belt if the channel is dropped